backend/pyproject.toml
[tool.poetry]
name = "search-it-backend"
version = "0.1.0"
description = ""
authors = ["Jie Han <jiehan1029@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.11"
pdfplumber = "^0.11.5"
pydantic = "^2.10.6"
python-docx = "^1.1.2"
uvicorn = "^0.34.0"
haystack-ai = "^2.11.2"
markdown-it-py = "^3.0.0"
mdit-plain = "^1.0.1"
pypdf = "^5.4.0"
sentence-transformers = ">=3.0.0"
torch = "^2.6.0"
python-dotenv = "^1.1.0"
beanie = "^1.29.0"
motor = "^3.7.0"
qdrant-haystack = "^9.1.1"
google-ai-haystack = "^5.1.0"
ollama-haystack = "^2.3.0"
celery = "^5.5.0"
redis = "^5.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
